fix: set proper ShellCheck source for utils.bash
The ShellCheck directive currently points to `../lib/utils.bash`.
However, this raises an info log, as ShellCheck is not able to open the
file. As the linter runs from the top level of the repository, the right
path should be `./lib/utils.bash`.

In bin/install line 9:
source "${plugin_dir}/lib/utils.bash"
       ^----------------------------^ SC1091 (info): Not following:
       ../lib/utils.bash: openBinaryFile: does not exist (No such file
       or directory)